% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albopop-package.R
\docType{package}
\name{albopop-package}
\alias{albopop-package}
\title{Read 'Albo Pretorio' pages and build a RSS-Feed}
\description{
Provides functions to read 'Albo Pretorio' pages and export a RSS-Feed
}
\author{
Patrick Hausmann \email{patrick.hausmann@covimo.de}
}

