% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_build_msg.R
\name{ap_build_msg}
\alias{ap_build_msg}
\title{Build Telegram and Twitter feeds}
\usage{
ap_build_msg(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A A dataframe
}
\description{
This takes data.frame of type 'albopop' and export a XML file
}

