% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_parse_page.R
\name{ap_parse_page}
\alias{ap_parse_page}
\title{Fetch the data.}
\usage{
ap_parse_page(url, link_url = NULL, site = "studiok")
}
\arguments{
\item{url}{A URL}

\item{link_url}{The URL for the links, if NULL same as `url`}

\item{site}{character, sets the input, currently four sites are supported: "JCityGov", "mapweb", "saga", "studiok"}
}
\value{
A dataframe
}
\description{
This scrapes the 'Albo Pretorio' data
}

