% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_send_telegram.R
\name{ap_send_telegram}
\alias{ap_send_telegram}
\title{Send a message via Telegram}
\usage{
ap_send_telegram(text_msg, TELEGRAM_TOKEN, TELEGRAM_CHAT_ID)
}
\arguments{
\item{text_msg}{character Message to send}

\item{TELEGRAM_TOKEN}{character Telegram Token}

\item{TELEGRAM_CHAT_ID}{character Telegram Chat_id}
}
\description{
Send a message via Telegram
}

