% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{disk}
\alias{disk}
\title{Options for saving Albopop feeds.}
\usage{
disk(path = "feed", overwrite = TRUE)
}
\arguments{
\item{path}{Path to store files in. A directory, not a file. Default: \code{feed}}

\item{overwrite}{(logical) Overwrite an existing file of the same name?
Default: \code{TRUE}}
}
\description{
Options for saving Albopop feeds.
}

