% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Set_Date}
\alias{Set_Date}
\title{Convert a Date in 'C' locale.}
\usage{
Set_Date(date)
}
\arguments{
\item{date}{a date string}
}
\value{
A date
}
\description{
This takes a date character and convert it to a 'C' local date.
}

